Parse individual fields with aeson
