+++ csv.cabal
@@ -18,6 +18,6 @@ Author:              Jaap Weel <weel@ugcs.caltech.edu>
 Maintainer:          Jaap Weel <weel@ugcs.caltech.edu>
 
 Library
-    Build-Depends:       base, parsec, filepath
+    Build-Depends:       base, parsec, filepath, aeson, scientific, text, utf8-string
     build-depends:       base >= 2 && < 5
-    Exposed-modules:     Text.CSV
+    Exposed-modules:     Text.CSV.Aeson
